+++ src/client/App.tsx
@@ -56,10 +56,10 @@ import ErfpachtFacturen from './pages/Erfpacht/ErfpachtFacturen';
 import ErfpachtOpenFacturen from './pages/Erfpacht/ErfpachtOpenFacturen';
 import GarbageInformation from './pages/GarbageInformation/GarbageInformation';
 import GeneralInfo from './pages/GeneralInfo/GeneralInfo';
-import HLI from './pages/HLI/HLI';
+import HLI from './pages/HLI/HLIThemaPagina';
 import HLIRegeling from './pages/HLI/HLIRegeling';
 import HLIRegelingen from './pages/HLI/HLIRegelingen';
-import HLIStadspasDetail from './pages/HLI/HLIStadspas';
+import HLIStadspasDetail from './pages/HLI/HLIStadspasDetail';
 import Horeca from './pages/Horeca/Horeca';
 import HorecaDetail from './pages/HorecaDetail/HorecaDetail';
 import Inkomen from './pages/Inkomen/Inkomen';
